requirements.txt
numpy>=1.19.4
opencv_python>=4.4.0.46
pyserial>=3.5
